# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zuultimate
  version: 1.0.0
paths:
...
  /v1/privacy/dsar:
    post:
      tags:
      - privacy
      summary: Submit a new DSAR
      operationId: submit_dsar_v1_privacy_dsar_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DSARSubmitRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DSARResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: beta
      x-required-scopes:
      - identity:write
    get:
      tags:
      - privacy
      summary: List DSAR requests
      operationId: list_dsars_v1_privacy_dsar_get
      security:
      - HTTPBearer: []
      parameters:
      - name: tenant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      - name: subject_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DSARResponse'
                title: Response List Dsars V1 Privacy Dsar Get
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: beta
      x-required-scopes:
      - identity:read
  /v1/privacy/dsar/{dsar_id}:
    get:
      tags:
      - privacy
      summary: Get a DSAR by ID
      operationId: get_dsar_v1_privacy_dsar__dsar_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: dsar_id
        in: path
        required: true
        schema:
          type: string
          title: Dsar Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DSARResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: beta
      x-required-scopes:
      - identity:read
  /v1/privacy/dsar/{dsar_id}/advance:
    post:
      tags:
      - privacy
      summary: Advance DSAR status
      operationId: advance_dsar_v1_privacy_dsar__dsar_id__advance_post
      security:
      - HTTPBearer: []
      parameters:
      - name: dsar_id
        in: path
        required: true
        schema:
          type: string
          title: Dsar Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DSARAdvanceRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DSARResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: beta
      x-required-scopes:
      - identity:write
  /v1/privacy/dsar/{dsar_id}/evidence:
    get:
      tags:
      - privacy
      summary: Get DSAR evidence pack
      operationId: get_evidence_pack_v1_privacy_dsar__dsar_id__evidence_get
      security:
      - HTTPBearer: []
      parameters:
      - name: dsar_id
        in: path
        required: true
        schema:
          type: string
          title: Dsar Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: beta
      x-required-scopes:
      - identity:read
components:
  schemas:
...
    DSARAdvanceRequest:
      properties:
        status:
          type: string
          pattern: ^(validated|processing|fulfilled|rejected)$
          title: Status
      type: object
      required:
      - status
      title: DSARAdvanceRequest
    DSARResponse:
      properties:
        id:
          type: string
          title: Id
        tenant_id:
          type: string
          title: Tenant Id
        subject_id:
          type: string
          title: Subject Id
        request_type:
          type: string
          title: Request Type
        status:
          type: string
          title: Status
        received_at:
          type: string
          title: Received At
        due_at:
          type: string
          title: Due At
        fulfilled_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Fulfilled At
        evidence_trail:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Evidence Trail
        is_overdue:
          type: boolean
          title: Is Overdue
          default: false
      type: object
      required:
      - id
      - tenant_id
      - subject_id
      - request_type
      - status
      - received_at
      - due_at
      - evidence_trail
      title: DSARResponse
    DSARSubmitRequest:
      properties:
        tenant_id:
          type: string
          title: Tenant Id
        subject_id:
          type: string
          title: Subject Id
        request_type:
          type: string
          pattern: ^(access|deletion|portability|correction|restriction)$
          title: Request Type
      type: object
      required:
      - tenant_id
      - subject_id
      - request_type
      title: DSARSubmitRequest
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
          description: Error message
        code:
          type: string
          title: Code
          description: Machine-readable error code
          default: ZUUL_ERROR
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
          description: Additional error context
      type: object
      required:
      - error
      title: ErrorResponse
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer